add rust fmt config
