+++ src/rust/rustfmt.toml
@@ -1,0 +1,3 @@
+imports_granularity = "Crate"
+group_imports = "StdExternalCrate"
+ignore = ["dependencies"]
